ref: modified helper schemas like dateString, stringToNumber
The reason for this change is that when properties using these schemas transformed from a string to a number, then re-validating the same object would violate the schema. That is why, the accepted types of these utility schemas were adjusted to accept e.g. strings, null and numbers to then transfer them to a number or null only.