feat: added variations collection to getProducts
When fetching products all variations of variable products are fetched as well. The reason for this is that the parent product of variable products does not bring a lot of value as it holds almost no information on its variations. That is why, the variations are fetched and added as products by merging the variable-parent and variation and by then removing the parent-variable product.